requirements.txt
streamlit
google-genai
pandas
sqlite3
python-dotenv
